moved tx set rebate
